+++ shuttle-engine/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "shuttle-runtime"
+name = "shuttle-engine"
 version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
