+++ shuttle-engine/src/hint.rs
