+++ shuttle-schedulers/src/random.rs
@@ -2,11 +2,11 @@ use rand::rngs::OsRng;
 use rand::seq::SliceRandom;
 use rand::{RngCore, SeedableRng};
 use rand_pcg::Pcg64Mcg;
-use shuttle_runtime::runtime::task::{Task, TaskId};
-use shuttle_runtime::scheduler::data::random::RandomDataSource;
-use shuttle_runtime::scheduler::data::DataSource;
-use shuttle_runtime::scheduler::{Schedule, Scheduler};
-use shuttle_runtime::seed_from_env;
+use shuttle_engine::runtime::task::{Task, TaskId};
+use shuttle_engine::scheduler::data::random::RandomDataSource;
+use shuttle_engine::scheduler::data::DataSource;
+use shuttle_engine::scheduler::{Schedule, Scheduler};
+use shuttle_engine::seed_from_env;
 
 /// A scheduler that randomly chooses a runnable task at each context switch.
 ///
