+++ shuttle-schedulers/src/round_robin.rs
@@ -1,7 +1,7 @@
-use shuttle_runtime::runtime::task::{Task, TaskId};
-use shuttle_runtime::scheduler::data::random::RandomDataSource;
-use shuttle_runtime::scheduler::data::DataSource;
-use shuttle_runtime::scheduler::{Schedule, Scheduler};
+use shuttle_engine::runtime::task::{Task, TaskId};
+use shuttle_engine::scheduler::data::random::RandomDataSource;
+use shuttle_engine::scheduler::data::DataSource;
+use shuttle_engine::scheduler::{Schedule, Scheduler};
 
 /// A round robin scheduler that chooses the next available runnable task at each context switch.
 #[derive(Debug)]
