+++ shuttle-schedulers/src/uncontrolled_nondeterminism.rs
@@ -1,5 +1,5 @@
-use shuttle_runtime::runtime::task::{Task, TaskId, DEFAULT_INLINE_TASKS};
-use shuttle_runtime::scheduler::{Schedule, Scheduler};
+use shuttle_engine::runtime::task::{Task, TaskId, DEFAULT_INLINE_TASKS};
+use shuttle_engine::scheduler::{Schedule, Scheduler};
 use smallvec::SmallVec;
 
 #[derive(Clone, Debug, PartialEq, Eq)]
