+++ shuttle-std/src/future.rs
@@ -5,9 +5,9 @@
 //!
 //! [`futures::executor`]: https://docs.rs/futures/0.3.13/futures/executor/index.html
 
-use shuttle_runtime::runtime::execution::ExecutionState;
-use shuttle_runtime::runtime::task::TaskId;
-use shuttle_runtime::runtime::thread;
+use shuttle_engine::runtime::execution::ExecutionState;
+use shuttle_engine::runtime::task::TaskId;
+use shuttle_engine::runtime::thread;
 use std::error::Error;
 use std::fmt::{Display, Formatter};
 use std::future::Future;
@@ -18,7 +18,7 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::task::{Context, Poll, Waker};
 
-pub use shuttle_runtime::future::batch_semaphore;
+pub use shuttle_engine::future::batch_semaphore;
 
 fn spawn_inner<F>(fut: F, caller: &'static Location<'static>) -> JoinHandle<F::Output>
 where
