+++ shuttle-std/src/sync/barrier.rs
@@ -1,8 +1,8 @@
 use crate::sync::{ResourceSignature, ResourceType};
-use shuttle_runtime::runtime::execution::ExecutionState;
-use shuttle_runtime::runtime::task::clock::VectorClock;
-use shuttle_runtime::runtime::task::TaskId;
-use shuttle_runtime::runtime::thread;
+use shuttle_engine::runtime::execution::ExecutionState;
+use shuttle_engine::runtime::task::clock::VectorClock;
+use shuttle_engine::runtime::task::TaskId;
+use shuttle_engine::runtime::thread;
 use std::cell::RefCell;
 use std::collections::HashSet;
 use std::fmt;
