+++ shuttle/Cargo.toml
@@ -10,7 +10,7 @@ categories = ["asynchronous", "concurrency", "development-tools::testing"]
 readme = "../README.md"
 
 [dependencies]
-shuttle-runtime = { path = "../shuttle-runtime" }
+shuttle-engine = { path = "../shuttle-engine" }
 shuttle-schedulers = { path = "../shuttle-schedulers" }
 shuttle-std = { path = "../shuttle-std" }
 bitvec = "1.0.1"
@@ -50,13 +50,13 @@ bench = false
 
 [features]
 default = []
-vector-clocks = ["shuttle-runtime/vector-clocks"]
-annotation = ["shuttle-runtime/annotation", "shuttle-schedulers/annotation", "dep:serde", "dep:serde_json", "dep:regex"]
+vector-clocks = ["shuttle-engine/vector-clocks"]
+annotation = ["shuttle-engine/annotation", "shuttle-schedulers/annotation", "dep:serde", "dep:serde_json", "dep:regex"]
 # The following feature overrides the `vector-clocks` feature. It SHOULD NOT be used in production.
 # Its purpose is solely to allow `cargo bench` to be run without vector clocks enabled, as they
 # are otherwise always enabled via a dev-dependency to ensure all *test* assertions utilizing vector
 # clocks behave correctly during testing
-bench-no-vector-clocks = ["shuttle-runtime/bench-no-vector-clocks"]
+bench-no-vector-clocks = ["shuttle-engine/bench-no-vector-clocks"]
 
 [[bench]]
 name = "lock"
