+++ shuttle/src/scheduler/mod.rs
@@ -1,7 +1,7 @@
 //! Implementations of different scheduling strategies for concurrency testing.
 
-// Re-export core scheduler types from shuttle-runtime
-pub use shuttle_runtime::scheduler::{DataSource, RandomDataSource, Schedule, Scheduler, Task, TaskId};
+// Re-export core scheduler types from shuttle-engine
+pub use shuttle_engine::scheduler::{DataSource, RandomDataSource, Schedule, Scheduler, Task, TaskId};
 
 // Re-export scheduler implementations from shuttle-schedulers
 pub use shuttle_schedulers::{
